feat(todo-intent): show worker activity tab in unified panel
