+++ free-todo-frontend/lib/i18n/messages/en.json
@@ -900,10 +900,28 @@
 		"clear": "Clear",
 		"expand": "Expand",
 		"collapse": "Collapse",
+		"tabRecords": "Records",
+		"tabWorkers": "Workers",
 		"connected": "Connected",
 		"disconnected": "Disconnected",
 		"reconnecting": "Reconnecting...",
 		"noRecords": "No intent records yet",
+		"workersOverview": "Worker overview",
+		"workersList": "Worker list",
+		"statusLoading": "Loading status...",
+		"statusUnavailable": "Status unavailable",
+		"refreshStatus": "Refresh status",
+		"configuredWorkers": "Configured workers",
+		"runningWorkers": "Running workers",
+		"activeWorkers": "Active workers",
+		"eventQueue": "Event queue",
+		"contextQueue": "Context queue",
+		"processedTotal": "Processed events",
+		"droppedTotal": "Dropped events",
+		"workerState": "State",
+		"workerActive": "Active",
+		"workerIdle": "Idle",
+		"workerOffline": "Offline",
 		"app": "App",
 		"window": "Window",
 		"speaker": "Speaker",
