+++ free-todo-frontend/lib/i18n/messages/zh.json
@@ -900,10 +900,28 @@
 		"clear": "清空",
 		"expand": "展开",
 		"collapse": "收起",
+		"tabRecords": "记录流",
+		"tabWorkers": "Worker",
 		"connected": "已连接",
 		"disconnected": "未连接",
 		"reconnecting": "重连中...",
 		"noRecords": "暂无识别记录",
+		"workersOverview": "Worker 概览",
+		"workersList": "Worker 列表",
+		"statusLoading": "状态加载中...",
+		"statusUnavailable": "状态暂不可用",
+		"refreshStatus": "刷新状态",
+		"configuredWorkers": "配置 Worker",
+		"runningWorkers": "运行中 Worker",
+		"activeWorkers": "工作中 Worker",
+		"eventQueue": "事件队列",
+		"contextQueue": "上下文队列",
+		"processedTotal": "已处理事件",
+		"droppedTotal": "已丢弃事件",
+		"workerState": "状态",
+		"workerActive": "工作中",
+		"workerIdle": "空闲",
+		"workerOffline": "离线",
 		"app": "应用",
 		"window": "窗口",
 		"speaker": "说话人",
